Add files via upload
Fifa 19 EDA is a GitHub project that offers exploratory data analysis (EDA) on the popular video game, FIFA 19. With this project, users can access insightful visualizations, statistical summaries, and data-driven insights to gain a better understanding of the game's player attributes, performance metrics, and team dynamics. It provides a comprehensive analysis of the game's data, enhancing the gaming experience and fostering data-driven decision-making.